requirements.txt
mega.py==1.0.8